Add pixel depth to image format
The image format now holds both the pixel depth and the color type.
The pixel depth is closely tied to the Intensity trait. Each pixel type
should define its pixel depth.

Removed the MissingChannelData error from the list of
DecompressionErrors. We will re-add it in the context of deserializing
CompressedImages.