Increase MAX_CONTEXT for every intensity
We are now using the YCoCg color-space as an intermediate representation
of an image. For RGB signals with depth n, the depth of the Y signal
will be n, and the depth of Co and Cg will be n+1.
Therefore, the MAX_CONTEXT for u8 and u16 intensities will double.
Also add tests to check if all possible contexts are less than
MAX_CONTEXT.